feat: switch to etcd for server
NOTE: Still need to add the other etcd ha servers to the cluster. This
requires removing the existing nodes and re-adding them as servers
with a server address and token file.

This switch lead to a consistent 20% decrease in CPU usage for the main
server, and stopped a crash loop where it got overloaded by requests.