fix: disable dhcp on jables eth
The ethernet cable got disconnected and reconnected, resulting in a dhcp
assigned ip. However, this ip is not recognized by the etcd cluster,
resulting in crashes when attempting to start k3s. Really, since we have
ips in our network config, we could disable dhcp for all ethernet
devices on all machines.